feat: add start&stop analytics task

+++ pkg/task/streamshell.go
@@ -1,37 +1,53 @@
 package task
 
 import (
+	"errors"
 	"os/exec"
+	"sync"
+
+	"github.com/sirupsen/logrus"
 )
 
 type StreamShell struct {
 	Shell           string
 	Cmd             *exec.Cmd
-	PushMessageFlag bool
+	Stopped bool
 }
+var mu sync.RWMutex
 
-var PipeShellMap map[string]*StreamShell
+var PipeShellMap map[string]*StreamShell 
 
 func RunStreamShell(id string, shell string, rpcSend func(s string) error) error {
 	cmd := exec.Command("bash", "-c", shell)
 	pipeShell := &StreamShell{
 		Shell:           shell,
 		Cmd:             cmd,
-		PushMessageFlag: true,
+		Stopped: false,
 	}
+	mu.Lock()
 	PipeShellMap[id] = pipeShell
-	err := cmd.Start()
-	if err != nil {
-		return err
-	}
+	mu.Unlock()
+	defer ClearStreamShell(id)
+
 	reader, err := cmd.StdoutPipe()
 	if err != nil {
 		return err
 	}
+	err = cmd.Start()
+	if err != nil {
+		return err
+	}
 	buf := make([]byte, 1024)
+	
 	for {
 		n, err := reader.Read(buf)
-		if err != nil || !pipeShell.PushMessageFlag {
+		if err != nil {
+			mu.RLock()
+			stopped := PipeShellMap[id].Stopped
+			mu.RUnlock()
+			if stopped {
+				return errors.New("stop stream shell")
+			}
 			return nil
 		}
 		if n > 0 {
@@ -44,8 +60,29 @@ func RunStreamShell(id string, shell string, rpcSend func(s string) error) error
 	}
 }
 
-func StopStreamShell(id string)error {
-	PipeShellMap[id].PushMessageFlag = false
+func StopStreamShell(id string) error {
+	mu.Lock()
+	defer mu.Unlock()
+
+	if PipeShellMap[id] == nil {
+		return nil
+	}
+	PipeShellMap[id].Stopped = true
+
 	cmd := PipeShellMap[id].Cmd
+	if cmd.ProcessState!=nil && cmd.ProcessState.Exited() {
+		logrus.Infof("stream shell %s already stopped", id)
+		return nil
+	}
 	return cmd.Process.Kill()
+}
+
+func ClearStreamShell(id string)error{
+	mu.Lock()
+	defer mu.Unlock()
+	if PipeShellMap[id] == nil {
+		return nil
+	}
+	delete(PipeShellMap, id)
+	return nil
 }
